requirements/backend.txt
beautifulsoup4==4.12.3
fastapi==0.115.6
httpx==0.28.1
openai==1.58.1
pydantic==2.10.4
pydantic-settings==2.7.0
python-dotenv==1.0.1
python-multipart==0.0.20
uvicorn==0.24.0